+++ scripts/delete_old_data.py
@@ -3,4 +3,4 @@ from processor.database.stories_db import delete_old_stories
 
 Session = database.get_session_maker()
 with Session() as session:
-    delete_old_stories(session)
+    delete_old_stories(session, 30)