+++ scripts/queue_wayback_stories.py
@@ -14,10 +14,12 @@ import processor
 
 processor.disable_package_loggers()
 
+import mcmetadata.urls as urls
 from mc_providers.onlinenews import OnlineNewsWaybackMachineProvider
 
 import processor.database as database
 import processor.database.projects_db as projects_db
+import processor.database.stories_db as stories_db
 import processor.fetcher as fetcher
 import processor.mcdirectory as mcdirectory
 import processor.projects as projects
@@ -27,9 +29,9 @@ from processor.classifiers import download_models
 POOL_SIZE = 8  # used for fetching project domains and listing stories in parallel
 DEFAULT_DAY_OFFSET = 4  # stories don't get processed for a few days
 DEFAULT_DAY_WINDOW = 3  # don't look for stories too old (DEFAULT_DAY_OFFSET + DEFAULT_DAY_WINDOW at most)
-PAGE_SIZE = 1000
+PAGE_SIZE = 200
 MAX_STORIES_PER_PROJECT = (
-    5000  # we can't process all the stories for queries that are too big
+    500  # we can't process all the stories for queries that are too big
 )
 
 
@@ -41,7 +43,7 @@ def load_projects() -> List[Dict]:
         force_reload=True, overwrite_last_story=False
     )
     logger.info("  Found {} projects".format(len(project_list)))
-    # return project_list[20:25]
+    # return project_list[20:22]
     # return [p for p in project_list if p['id'] == 166]
     return project_list
 
@@ -62,6 +64,7 @@ def _project_story_worker(p: Dict) -> List[Dict]:
         False,
     )
     project_stories = []
+    skipped_dupes = 0  # how many URLs do we filter out because they're already in the DB for this project recently
     page_number = 1
     full_project_query = _query_builder(p["search_terms"], p["language"])
     try:
@@ -76,6 +79,10 @@ def _project_story_worker(p: Dict) -> List[Dict]:
             )
         )
         if total_hits > 0:  # don't bother querying if no results to page through
+            # list recent urls to filter so we don't fetch text extra if we've recently proceses already (and will be filtered
+            # out by add_stories call in later post-text-fetch step)
+            with Session() as session:
+                already_processed_urls = stories_db.project_story_urls(session, p, 14)
             # using the provider wrapper so this does the chunking into smaller queries for us
             latest_pub_date = dt.datetime.now() - dt.timedelta(
                 weeks=50
@@ -102,6 +109,10 @@ def _project_story_worker(p: Dict) -> List[Dict]:
                 for item in page:
                     if len(project_stories) > MAX_STORIES_PER_PROJECT:
                         break
+                    # skip URLs we've processed recently
+                    if urls.normalize_url(item["url"]) in already_processed_urls:
+                        skipped_dupes += 1
+                        continue
                     info = dict(
                         id=item[
                             "id"
@@ -125,8 +136,8 @@ def _project_story_worker(p: Dict) -> List[Dict]:
                     project_stories.append(info)
                 time.sleep(0.1)  # don't hammer the API
             logger.info(
-                "  project {} - {} stories (after {})".format(
-                    p["id"], len(project_stories), start_date
+                "  project {} - {} stories (skipped {}) (after {})".format(
+                    p["id"], len(project_stories), skipped_dupes, start_date
                 )
             )
             # after all pages done, update latest pub date so we start at that next time
@@ -134,13 +145,13 @@ def _project_story_worker(p: Dict) -> List[Dict]:
                 projects_db.update_history(
                     session, p["id"], latest_pub_date, processor.SOURCE_WAYBACK_MACHINE
                 )
-    except Exception as e:
+    except Exception:
         # perhaps a query syntax error? log it, but keep going so other projects succeed
         logger.error(
             f"  project {p['id']} - failed to fetch stories (likely a query syntax or connection error)"
         )
         # logger.exception(e) #Ignore Sentry Logging
-    return project_stories
+    return project_stories[:MAX_STORIES_PER_PROJECT]
 
 
 def fetch_project_stories(project_list: List[Dict]) -> List[Dict]:
@@ -188,8 +199,8 @@ def fetch_text(stories: List[Dict]) -> List[Dict]:
 
     # download them all in parallel... will take a while (note that we're fetching the extracted content JSON here,
     # NOT the archived or original HTML because that saves us the parsing and extraction step)
-    urls = [s["extracted_content_url"] for s in stories]
-    fetcher.fetch_all_html(urls, handle_parse)
+    unique_urls = list(set([s["extracted_content_url"] for s in stories]))
+    fetcher.fetch_all_html(unique_urls, handle_parse)
     logger.info(
         "Fetched text for {} stories (failed on {})".format(
             len(stories_to_return), len(stories) - len(stories_to_return)
@@ -221,11 +232,20 @@ if __name__ == "__main__":
 
     # 3. fetch all the urls from for each project from wayback machine (serially so we don't have to flatten 😖)
     all_stories = fetch_project_stories(projects_with_domains)
-    logger.info("Discovered {} total story URLs".format(len(all_stories)))
+    unique_url_count = len(set([s["extracted_content_url"] for s in all_stories]))
+    logger.info(
+        "Discovered {} total stories, {} unique URLs".format(
+            len(all_stories), unique_url_count
+        )
+    )
 
     # 4. fetch pre-parsed content (will happen in parallel by story)
     stories_with_text = fetch_text(all_stories)
-    logger.info("Fetched {} stories with text".format(len(stories_with_text)))
+    logger.info(
+        "Fetched {} stories with text, from {} attempted URLs".format(
+            len(stories_with_text), unique_url_count
+        )
+    )
 
     # 5. post batches of stories for classification
     results_data = tasks.queue_stories_for_classification(